// frontend/generated/models/ProjectStatusDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Freelance Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ProjectStatusDto = {
    Open: 'OPEN',
    Paused: 'PAUSED',
    InProgress: 'IN_PROGRESS',
    Terminated: 'TERMINATED',
    Finished: 'FINISHED'
} as const;
export type ProjectStatusDto = typeof ProjectStatusDto[keyof typeof ProjectStatusDto];


export function instanceOfProjectStatusDto(value: any): boolean {
    for (const key in ProjectStatusDto) {
        if (Object.prototype.hasOwnProperty.call(ProjectStatusDto, key)) {
            if (ProjectStatusDto[key as keyof typeof ProjectStatusDto] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ProjectStatusDtoFromJSON(json: any): ProjectStatusDto {
    return ProjectStatusDtoFromJSONTyped(json, false);
}

export function ProjectStatusDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProjectStatusDto {
    return json as ProjectStatusDto;
}

export function ProjectStatusDtoToJSON(value?: ProjectStatusDto | null): any {
    return value as any;
}

export function ProjectStatusDtoToJSONTyped(value: any, ignoreDiscriminator: boolean): ProjectStatusDto {
    return value as ProjectStatusDto;
}